STANFORD_DB/requirements.txt
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-dotenv==1.0.0 